refactor: shift sumcheck [JOLT-227] (#1087)
* refactor: rename r_cycle to r_outer for consistency with outer sumcheck naming

- Renamed all occurrences of r_cycle to r_outer in shift.rs
- Updated variable names and comments to reflect the change
- This aligns with the naming convention that this is the outer sumcheck opening point

* refactor: address PR review comments on #1051

- Moved EqPlusOnePolynomial and EqPlusOnePrefixSuffixPoly to new file eq_plus_one_poly.rs
- Added prefix_suffix() method to EqPlusOnePolynomial for better API
- Moved CycleState struct to r1cs/inputs.rs where shift sumcheck inputs belong
- All tests passing including critical muldiv e2e test

* refactor: rename CycleState to ShiftSumcheckCycleState for clarity

- More descriptive name that clearly indicates its purpose for shift sumcheck
- Updated all references in both inputs.rs and shift.rs

* refactor: do not recalculate same eq evals


* fix tests


---------
