+++ jolt-core/src/poly/eq_poly.rs
@@ -193,94 +193,6 @@ impl<F: JoltField> EqPolynomial<F> {
     }
 }
 
-/// Multilinear extension for the relation y = x + 1 over bitstrings (no wrap-around; big-endian inputs).
-pub struct EqPlusOnePolynomial<F: JoltField> {
-    x: Vec<F::Challenge>,
-}
-
-impl<F: JoltField> EqPlusOnePolynomial<F> {
-    pub fn new(x: Vec<F::Challenge>) -> Self {
-        EqPlusOnePolynomial { x }
-    }
-
-    /* This MLE is 1 if y = x + 1 for x in the range [0... 2^l-2].
-    That is, it ignores the case where x is all 1s, outputting 0.
-    Assumes x and y are provided big-endian. */
-    pub fn evaluate(&self, y: &[F::Challenge]) -> F {
-        let l = self.x.len();
-        let x = &self.x;
-        assert!(y.len() == l);
-        let one = F::from_u64(1_u64);
-
-        /* If y+1 = x, then the two bit vectors are of the following form.
-            Let k be the longest suffix of 1s in x.
-            In y, those k bits are 0.
-            Then, the next bit in x is 0 and the next bit in y is 1.
-            The remaining higher bits are the same in x and y.
-        */
-        (0..l)
-            .into_par_iter()
-            .map(|k| {
-                let lower_bits_product = (0..k)
-                    .map(|i| x[l - 1 - i] * (F::one() - y[l - 1 - i]))
-                    .product::<F>();
-                let kth_bit_product = (F::one() - x[l - 1 - k]) * y[l - 1 - k];
-                let higher_bits_product = ((k + 1)..l)
-                    .map(|i| {
-                        x[l - 1 - i] * y[l - 1 - i] + (one - x[l - 1 - i]) * (one - y[l - 1 - i])
-                    })
-                    .product::<F>();
-                lower_bits_product * kth_bit_product * higher_bits_product
-            })
-            .sum()
-    }
-
-    #[tracing::instrument(skip_all, "EqPlusOnePolynomial::evals")]
-    pub fn evals(r: &[F::Challenge], scaling_factor: Option<F>) -> (Vec<F>, Vec<F>) {
-        let ell = r.len();
-        let mut eq_evals: Vec<F> = unsafe_allocate_zero_vec(ell.pow2());
-        eq_evals[0] = scaling_factor.unwrap_or(F::one());
-        let mut eq_plus_one_evals: Vec<F> = unsafe_allocate_zero_vec(ell.pow2());
-
-        // i indicates the LENGTH of the prefix of r for which the eq_table is calculated
-        let eq_evals_helper = |eq_evals: &mut Vec<F>, r: &[F::Challenge], i: usize| {
-            debug_assert!(i != 0);
-            let step = 1 << (ell - i); // step = (full / size)/2
-
-            let mut selected: Vec<_> = eq_evals.par_iter_mut().step_by(step).collect();
-
-            selected.par_chunks_mut(2).for_each(|chunk| {
-                *chunk[1] = *chunk[0] * r[i - 1];
-                *chunk[0] -= *chunk[1];
-            });
-        };
-
-        for i in 0..ell {
-            let step = 1 << (ell - i);
-            let half_step = step / 2;
-
-            let mut r_lower_product = F::one();
-            for &x in r.iter().skip(i + 1) {
-                r_lower_product = r_lower_product * x; // To get the benefits of multiplication
-            }
-            r_lower_product *= F::one() - r[i];
-
-            eq_plus_one_evals
-                .par_iter_mut()
-                .enumerate()
-                .skip(half_step)
-                .step_by(step)
-                .for_each(|(index, v)| {
-                    *v = eq_evals[index - half_step] * r_lower_product;
-                });
-
-            eq_evals_helper(&mut eq_evals, r, i + 1);
-        }
-
-        (eq_evals, eq_plus_one_evals)
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
